Change X-UA-Device detect fallback to active theme
If the detect_device method sees an X-UA-Device header it will attempt to set
the theme accordingly. However if it does not find a handheld or tablet device
it will default to the mobile theme. This is not desirable as it means the
actual desktop theme is never used. The low device support has been removed
from the else condition (and firefoxos support added to the mobile group).